+++ src/map/unit.h
@@ -121,7 +121,7 @@ int unit_set_target(struct unit_data* ud, int target_id);
 // unit_data ‚Ì‰Šú‰»ˆ—
 void unit_dataset(struct block_list *bl);
 
-int unit_fixdamage(struct block_list *src,struct block_list *target,unsigned int tick,int sdelay,int ddelay,int damage,int div,int type,int damage2);
+int unit_fixdamage(struct block_list *src,struct block_list *target,unsigned int tick,int sdelay,int ddelay,int64 damage,int div,int type,int64 damage2);
 // ‚»‚Ì‘¼
 struct unit_data* unit_bl2ud(struct block_list *bl);
 struct unit_data* unit_bl2ud2(struct block_list *bl);
